Port code to support Python3 as well as Python2
As there's no tests, the judgement of this working is based off a) the module
importing cleaning & b) the example log consumer running.

This introduces a new dependancy on six, which is required to support meta
classes across the versions.

+++ src/openbmp/api/parsed/message/Message.py
@@ -28,8 +28,8 @@ class Message(object):
         self.version = float()
         self.type = str()
         self.collector_hash_id = str()
-        self.length = long()
-        self.records = long()
+        self.length = int()
+        self.records = int()
         self.router_hash_id = str()
         self.content = str()
         self.content_pos = int()
@@ -67,10 +67,10 @@ class Message(object):
                 self.type = value
 
             elif attr == "L":
-                self.length = long(value)
+                self.length = int(value)
 
             elif attr == "R":
-                self.records = long(value)
+                self.records = int(value)
 
             elif attr == "R_HASH_ID":
                 self.router_hash_id = value
